+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `declaration` property to XmlDocumentNode.
 - Add `recycleNodes` option to XmlDocumentNode.from() options.
 - Add cell recycling for use with combined tuning.
+- Add `fromRecycled()` method on XmlDocumentNode.
 ### Changed
 - Rename `includeProcessingInstructions` option in `XmlFormattingOptions` to `writeXmlDeclaration`.
 - XML documents now preserve the XML declaration they had when parsed.